requirements.txt
python-telegram-bot>=20.0
httpx>=0.24.0
telethon>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
mcp>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0